+++ src/main/kotlin/com/scrumdapp/groupservice/services/GroupFeatureService.kt
@@ -13,7 +13,7 @@ class GroupFeatureService(
 ) {
 
     fun create(dto: GroupFeatureDto): GroupFeatureDto {
-        val group = groupRepository.findById(dto.key!!)
+        val group = groupRepository.findById(dto.key!!.toInt())
             .orElseThrow { NotFoundException("Group with id ${dto.key} not found") }
 
         val entity = GroupFeatureMapper.toEntity(dto, group)
